+++ tests/integration/test_client_covid_symptoms.py
