+++ tests/resources/test_negation_hard.txt
